feat(auth): add full authentication system with email, password reset, and Gmail integration
