+++ src/app/reset-password/page.tsx
@@ -1,17 +1,10 @@
+import { Suspense } from 'react';
 import ResetPasswordForm from '@/components/auth/ResetPasswordForm';
 
 export default function ResetPasswordPage() {
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div className="text-center">
-          <h1 className="text-3xl font-bold text-gray-900">Project Manager</h1>
-          <p className="mt-2 text-sm text-gray-600">
-            Set your new password
-          </p>
-        </div>
-        <ResetPasswordForm />
-      </div>
-    </div>
+    <Suspense>
+      <ResetPasswordForm />
+    </Suspense>
   );
 }
