+++ src/app/verify-email/page.tsx
@@ -1,104 +1,10 @@
-'use client';
+import { Suspense } from 'react';
+import VerifyEmailPage from '../../components/auth/VerifyEmailPage';
 
-import { useEffect, useState } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Alert, AlertDescription } from '@/components/ui/alert';
-import { CheckCircle, AlertCircle, ArrowRight } from 'lucide-react';
-
-export default function VerifyEmailPage() {
-  const [status, setStatus] = useState<'loading' | 'success' | 'error'>('loading');
-  const [message, setMessage] = useState('');
-  const router = useRouter();
-  const searchParams = useSearchParams();
-
-  useEffect(() => {
-    const verifyEmail = async () => {
-      const token = searchParams.get('token');
-      
-      if (!token) {
-        setStatus('error');
-        setMessage('Verification token is missing');
-        return;
-      }
-
-      try {
-        const response = await fetch(`/api/auth/verify-email?token=${token}`);
-        const data = await response.json();
-
-        if (response.ok) {
-          setStatus('success');
-          setMessage(data.message);
-        } else {
-          setStatus('error');
-          setMessage(data.error || 'Email verification failed');
-        }
-              } catch {
-          setStatus('error');
-          setMessage('An error occurred during verification');
-        }
-    };
-
-    verifyEmail();
-  }, [searchParams]);
-
+export default function Page() {
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div className="text-center">
-          <h1 className="text-3xl font-bold text-gray-900">Project Manager</h1>
-          <p className="mt-2 text-sm text-gray-600">
-            Email Verification
-          </p>
-        </div>
-        
-        <Card className="w-full">
-          <CardHeader className="space-y-1">
-            <CardTitle className="text-2xl font-bold text-center">
-              {status === 'loading' && 'Verifying Email...'}
-              {status === 'success' && 'Email Verified!'}
-              {status === 'error' && 'Verification Failed'}
-            </CardTitle>
-            <CardDescription className="text-center">
-              {status === 'loading' && 'Please wait while we verify your email address'}
-              {status === 'success' && 'Your email has been successfully verified'}
-              {status === 'error' && 'There was an issue verifying your email'}
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            {status === 'loading' && (
-              <div className="text-center">
-                <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-gray-900 mx-auto"></div>
-              </div>
-            )}
-            
-            {status === 'success' && (
-              <Alert>
-                <CheckCircle className="h-4 w-4" />
-                <AlertDescription>{message}</AlertDescription>
-              </Alert>
-            )}
-            
-            {status === 'error' && (
-              <Alert variant="destructive">
-                <AlertCircle className="h-4 w-4" />
-                <AlertDescription>{message}</AlertDescription>
-              </Alert>
-            )}
-
-            <div className="text-center">
-              <Button
-                onClick={() => router.push('/login')}
-                className="w-full"
-              >
-                {status === 'success' ? 'Continue to Login' : 'Back to Login'}
-                <ArrowRight className="h-4 w-4 ml-2" />
-              </Button>
-            </div>
-          </CardContent>
-        </Card>
-      </div>
-    </div>
+    <Suspense>
+      <VerifyEmailPage />
+    </Suspense>
   );
 }
